HomalgProject.jl

Introduction

The Julia package HomalgProject provides simplified access to the repositories of the GAP packages hosted at the GitHub organization homalg-project, most of which are based on the

CAP project,

homalg project.

These are open source GAP multi-package projects for constructive category theory and homological algebra with applications to module theory of commutative and non-commutative algebras and algebraic geometry.

Installation

To install Julia follow the instructions of the first two steps on the Oscar installation page. Then start Julia and add the package HomalgProject:

$ julia
julia> using Pkg; Pkg.add("HomalgProject")

Testing

The correctness of the installation and the availability of the functionality can at any time be tested using

julia> using Pkg; Pkg.test("HomalgProject")

Updating

julia> using Pkg; Pkg.update("HomalgProject")
julia> Pkg.build("HomalgProject")

This will also clone the repositories listed in HomalgProject.PACKAGES_TO_DOWNLOAD using DownloadAllPackagesFromHomalgProject and compile the packages listed in HomalgProject.PACKAGES_TO_COMPILE using GAP.Packages.install(pkgname).

HomalgProject.version — Constant

HomalgProject.version

The version number of the loaded HomalgProject. Please mention this number in any bug report.

HomalgProject.PKG_DIR — Constant

HomalgProject.PKG_DIR

The directory in which to install bundled GAP packages from the github organization https://github.com/homalg-project/. It is equal to joinpath(pathof(HomalgProject),"pkg").

HomalgProject.PACKAGES_TO_DOWNLOAD — Constant

HomalgProject.PACKAGES_TO_DOWNLOAD

List of packages which will be considered by

DownloadAllPackagesFromHomalgProject()

UpdateAllPackagesFromHomalgProject()

RemoveAllPackagesFromHomalgProject()

HomalgProject.DownloadAllPackagesFromHomalgProject — Function

DownloadAllPackagesFromHomalgProject()

Apply DownloadPackageFromHomalgProject to all packages listed in PACKAGES_TO_DOWNLOAD.

HomalgProject.UpdateAllPackagesFromHomalgProject — Function

UpdateAllPackagesFromHomalgProject()

Apply UpdatePackageFromHomalgProject to all packages listed in PACKAGES_TO_DOWNLOAD.

HomalgProject.RemoveAllPackagesFromHomalgProject — Function

RemoveAllPackagesFromHomalgProject()

Apply RemovePackageFromHomalgProject to all packages listed in PACKAGES_TO_DOWNLOAD.

HomalgProject.DownloadPackageFromHomalgProject — Function

DownloadPackageFromHomalgProject(pkgname)

Clone the package named pkgname from the github organization https://github.com/homalg-project/ to the subdirectory HomalgProject.PKG_DIR. On success return true and on failure false.

HomalgProject.UpdatePackageFromHomalgProject — Function

UpdatePackageFromHomalgProject(pkgname)

Update the package named pkgname located in the subdirectory HomalgProject.PKG_DIR from the github organization https://github.com/homalg-project/. If the package directory does not exist locally then invoke DownloadPackageFromHomalgProject(pkgname). On success return true and on failure false.

HomalgProject.RemovePackageFromHomalgProject — Function

RemovePackageFromHomalgProject(pkgname)

Delete the package named pkgname from the subdirectory HomalgProject.PKG_DIR. On success return true and on failure false. Removing a package and re-downloading it might be useful if udpating it fails.

HomalgProject.PACKAGES_TO_COMPILE — Constant

HomalgProject.PACKAGES_TO_COMPILE

The list of all GAP packages that will downloaded (once) and installed by GAP.Packages.install when using Pkg; Pkg.build("HomalgProject") is invoked. The latter should be called once GAP.jl gets updated.

Software dependency

HomalgProject relies on the

computer algebra systems	through the Julia packages

GAP	Gap.jl

Nemo	Nemo = Nemo.jl

Singular	Singular.jl

all of which are components of the computer algebra system OSCAR.

Some of the bundled packages use the GAP packages

IO

ferret

json

QPA2

and the

third party software	through the GAP packages

cddlib	CddInterface

4ti2	4ti2Interface

Normaliz	NormalizInterface

Graphviz	Digraphs

General Disclaimer

The software comes with absolutely no warranty and will most likely have errors. If you use it for computations, please check the correctness of the result very carefully.

This software is licensed under the LGPL, version 3, or any later version.

Funding

The development of both software projects was partially funded by the DFG (German Research Foundation) through the Special Priority Project SPP 1489 and the Transregional Collaborative Research Centre SFB-TRR 195.

Index

HomalgProject.PACKAGES_TO_COMPILE

HomalgProject.PACKAGES_TO_DOWNLOAD

HomalgProject.PKG_DIR

HomalgProject.version

HomalgProject.DownloadAllPackagesFromHomalgProject

HomalgProject.DownloadPackageFromHomalgProject

HomalgProject.RemoveAllPackagesFromHomalgProject

HomalgProject.RemovePackageFromHomalgProject

HomalgProject.UpdateAllPackagesFromHomalgProject

HomalgProject.UpdatePackageFromHomalgProject